requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
flask-sqlalchemy
flask-migrate
psycopg2-binary
